# apps/gateway/openapi.yaml
openapi: 3.1.0
info:
  title: Mesh gateway
  version: 0.1.0
  summary: Trading fees become hourly AI inference credits, spent through an OpenAI-compatible gateway, served by Macs and frontier providers, and sold on a marketplace when unused.
  description: |
    Mesh is a token whose trading fees buy AI inference for the people holding it. A 1.5 % fee on every
    $MESH trade is swept once an hour; half goes to the network treasury and half is split, pro rata, across
    every wallet holding at least 1,000 tokens (time-weighted over the hour) as credits denominated in US
    dollars. A second engine, a share of the margin on paid requests and marketplace fees, pays into the same
    hourly pool (`GET /stats → usageShareEnabled`).
    Holders spend credits through this gateway with their own API keys: open models on Macs in the network at
    a flat price per million tokens, frontier models through zero-data-retention providers at list plus a 6 %
    markup that covers what the upstream charges Mesh (`GET /v1/models` carries both prices). Credits nobody
    will use are sold on the credit marketplace (`/market/*`, docs/MARKETPLACE.md), and anyone can buy credits
    from Mesh at face value with a prepaid balance (`/me/credits/buy`). A holder's first sign-in receives
    starter credits, which can be spent but not sold.

    What stands behind a credit, and what limits it:

    * **Reserve.** The holder share of every sweep is swapped to a stablecoin on chain and kept in a
...

...

...

...
servers:
...
security: []
paths:
...
  /v1/models:
    get:
      tags: [Inference]
      summary: The model catalogue, with the Mesh price next to list
      description: |
        The curated catalogue (docs/PRICING.md §1), not only what Mesh nodes run: the open models Macs in the
        network serve (`served: network | both`, also reachable by a short alias such as `llama-3.1-8b`) plus the
        frontier, fast and open models reached through the upstream with zero-data-retention providers
        (`served: upstream`). OpenAI shape (`object: list`, `data[].id / object / created / owned_by`) with
        per-model `displayName`, `vendor`, `tier`, `served`, `listPrice`, `meshPrice`, `privacy`, `online` and
        `guestAllowed`, and a top-level `pricing` block with the knobs from config. Works without a key (the web
        picker and the guest chat read it); a bearer that is present must be a valid key. `mesh_network` is kept
        for pre-catalogue clients.
      operationId: listModels
      security: [{ apiKey: [] }, {}]
      parameters:
        - in: query
          name: guest
          required: false
          schema: { type: string, enum: ['1'] }
          description: Narrow the list to what a homepage guest may pick (network models plus `config.guest.allowedTiers`).
      responses:
        '200':
          description: OpenAI-shaped list, network models first, then frontier, fast and open.
          content:
            application/json:
              schema:
                type: object
                required: [object, data, pricing]
                properties:
                  object: { type: string, const: list }
                  data: { type: array, items: { $ref: '#/components/schemas/Model' } }
                  pricing:
                    type: object
                    required: [networkPricePerMTokens, upstreamDiscountBps, upstreamMarkupBps, guestTiers]
                    properties:
                      networkPricePerMTokens: { type: number, description: Flat USD per 1M total tokens when a Mesh node serves the request. }
                      upstreamDiscountBps: { type: integer, description: Upstream models bill list × (1 − this / 10000). Exclusive with the markup. }
                      upstreamMarkupBps: { type: integer, description: Upstream models bill list × (1 + this / 10000). Exclusive with the discount. }
                      guestTiers: { type: array, items: { type: string, enum: [frontier, fast, open] }, description: Tiers a guest may use besides the network models. }
              example:
                object: list
                data:
                  - { id: meta-llama/llama-3.1-8b-instruct, object: model, created: 0, owned_by: mesh, name: Llama 3.1 8B, displayName: Llama 3.1 8B, vendor: Meta, tier: open, served: both, listPrice: { promptUsdPerM: 0.05, completionUsdPerM: 0.08 }, meshPrice: { promptUsdPerM: 0.08, completionUsdPerM: 0.08 }, privacy: network, online: 3, guestAllowed: true, mesh_network: true }
                  - { id: anthropic/claude-sonnet-4.5, object: model, created: 0, owned_by: anthropic, name: Claude Sonnet 4.5, displayName: Claude Sonnet 4.5, vendor: Anthropic, tier: frontier, served: upstream, listPrice: { promptUsdPerM: 3, completionUsdPerM: 15 }, meshPrice: { promptUsdPerM: 3.18, completionUsdPerM: 15.9 }, privacy: upstream_zdr, online: 0, guestAllowed: false, mesh_network: false }
                pricing: { networkPricePerMTokens: 0.08, upstreamDiscountBps: 0, upstreamMarkupBps: 600, guestTiers: [open, fast] }
        '401': { $ref: '#/components/responses/OpenAIUnauthorized' }
        '429': { $ref: '#/components/responses/OpenAIRateLimited' }
        '502': { $ref: '#/components/responses/OpenAIUpstream' }
  /v1/chat/completions:
    post:
      tags: [Inference]
      summary: Chat completion
      description: |
        Same request and response shape as OpenAI. Set `stream: true` for server-sent events.

        * Privacy tier (docs/PRIVACY.md): `X-Mesh-Privacy: trusted | network | upstream_zdr`, or `mesh.privacy` in the body,
          else the key's default, else the gateway default (`trusted`). `trusted` = your own Macs, allowlisted operators, and
          operators with the required stake who signed the pledge; if none is online the request goes to the upstream with
          zero-data-retention providers only, never to another node. `network` = any eligible node. `upstream_zdr` = skip the network.
          The machine that runs the model sees the prompt in plaintext whichever tier you pick. Unknown value → `400 invalid_privacy_tier`.
        * When an idle Mesh node of the right tier advertises the model, it serves the request and you pay the flat network price
          ($0.08 per million total tokens). The response header `x-mesh-route` is `node:<nodeId>`.
        * When every such node is online but at capacity, the job is queued behind them for up to `routing.queueWaitMs`
          (6 s by default) and served by the first node that frees up (`mesh.routeReason: queued_then_node`). If none does,
          the request goes upstream with `x-mesh-fallback: queue_timeout`; when the queue is already `maxQueueDepthPerNode`
          deep per node it goes upstream at once with `x-mesh-fallback: queue_full`.
        * Otherwise the request goes to the upstream provider at list plus the configured markup (`GET /v1/models → pricing`;
          6 % as shipped, which covers the 5.5 % the upstream charges Mesh) and `x-mesh-route` names the upstream.
          If a node failed before any output reached you (or no trusted node was online), `x-mesh-fallback` carries the reason.
        * Every reply says how it was served: `x-mesh-privacy`, `x-mesh-served-by` and `mesh.privacy` / `mesh.servedBy`
          (`your node`, `trusted node`, `network node`, `upstream (ZDR)`, `upstream`) in the final chunk or the JSON body.
        * Non-streamed replies set `x-mesh-cost-usd` and `x-mesh-balance-usd`. Streamed replies carry `usage.cost`
          (USD) in the final chunk, plus `mesh: { route, routeReason, nodeId, privacy, servedBy, listCostUsd, savedUsd }` when a node served it
          (`routeReason` is `node`, `queued_then_node`, or for upstream replies why: `not_network_model`, `no_online_node`,
          `no_trusted_node`, `queue_timeout`, `queue_full`, `node_failed`, `upstream_requested`, `network_disabled`).
        * Failed requests are never charged.
        * Rate limit (`V1_RATE_LIMIT`, default 120 requests per minute) is counted **per API key** (per client IP for
          unauthenticated calls), with a fixed one-minute window, across `/v1/models` and `/v1/chat/completions` together.
          `x-ratelimit-limit` / `x-ratelimit-remaining` are on every reply and `retry-after` on the 429. A single key
          driving a burst of more than 120 requests a minute gets `429 rate_limited` after the 120th regardless of how
          many nodes are online: spread load over several keys (one per app or worker) or ask the operator to raise
          `V1_RATE_LIMIT`. Streams count once, when they start.
      operationId: createChatCompletion
      security: [{ apiKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChatRequest' }
            examples:
              simple:
                summary: One message, no stream
                value:
                  model: meta-llama/llama-3.1-8b-instruct
                  messages: [{ role: user, content: Summarise this clause. }]
              stream:
                summary: Streaming with usage
                value:
                  model: meta-llama/llama-3.1-8b-instruct
                  messages: [{ role: user, content: Summarise this clause. }]
                  stream: true
                  stream_options: { include_usage: true }
      responses:
        '200':
          description: Completion (JSON) or `text/event-stream` when `stream` is true.
          headers:
            x-mesh-route: { description: '`node:<nodeId>` or the upstream name.', schema: { type: string } }
            x-mesh-cost-usd: { description: 'Charged for this request (non-stream only).', schema: { type: string } }
            x-mesh-balance-usd: { description: 'Balance after this request (non-stream only).', schema: { type: string } }
            x-mesh-saved-usd: { description: 'List price minus what you paid when a node served it (non-stream only).', schema: { type: string } }
            x-mesh-fallback: { description: 'Reason a node attempt was abandoned before output (`unclaimed`, `first_token_timeout`, `node_error: …`, `queue_timeout`), or why none was tried (`no_trusted_node`, `queue_full`).', schema: { type: string } }
            x-mesh-privacy: { description: 'Tier the reply was served under.', schema: { $ref: '#/components/schemas/PrivacyTier' } }
            x-mesh-served-by: { description: '`your node` (a trusted request served by your own Mac), `trusted node`, `network node`, `upstream (ZDR)` or `upstream`.', schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ChatCompletion' }
              example:
                id: chatcmpl-mesh-m1x2y3
                object: chat.completion
                created: 1759487020
                model: meta-llama/llama-3.1-8b-instruct
                choices:
                  - index: 0
                    message: { role: assistant, content: "Either side can end the agreement with 30 days' notice, but fees already invoiced stay payable." }
                    finish_reason: stop
                usage: { prompt_tokens: 212, completion_tokens: 31, total_tokens: 243, cost: 0.0000049 }
                mesh: { route: node, nodeId: node_7kd2pq9f, chip: M2 Pro, privacy: trusted, servedBy: trusted node, listCostUsd: 0.000039, savedUsd: 0.000034 }
            text/event-stream:
              schema: { type: string }
              example: |
                data: {"id":"chatcmpl-mesh-m1x2y3","object":"chat.completion.chunk","created":1759487020,"model":"meta-llama/llama-3.1-8b-instruct","choices":[{"index":0,"delta":{"role":"assistant","content":"Either"},"finish_reason":null}]}

                data: {"id":"chatcmpl-mesh-m1x2y3","object":"chat.completion.chunk","created":1759487020,"model":"meta-llama/llama-3.1-8b-instruct","choices":[{"index":0,"delta":{},"finish_reason":"stop"}],"usage":{"prompt_tokens":212,"completion_tokens":31,"total_tokens":243,"cost":0.0000049},"mesh":{"route":"node","nodeId":"node_7kd2pq9f","listCostUsd":0.000039,"savedUsd":0.000034}}

                data: [DONE]
        '400':
          description: '`messages` empty or `model` missing.'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OpenAIError' }
        '401': { $ref: '#/components/responses/OpenAIUnauthorized' }
        '402':
          description: '`insufficient_quota`: credit balance is zero (credit past its 90 days has lapsed and does not count). Hold $MESH through an hour to receive credits, or buy some with `POST /me/credits/buy`.'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OpenAIError' }
              example: { error: { message: 'Insufficient Mesh credits (balance $0.000000). Hold $MESH to receive hourly credits, or buy credits on the market page.', type: insufficient_quota, code: insufficient_quota, param: null } }
        '403':
          description: '`model_not_allowed`: not in the model policy. `GET /v1/models` lists what is.'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OpenAIError' }
        '429':
          description: Per-key rate limit, or `key_spend_limit_reached` when the key's lifetime cap is used up.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OpenAIError' }
              example: { error: { message: 'This API key reached its spend limit ($5). Raise it with PATCH /keys/3.', type: insufficient_quota, code: key_spend_limit_reached, param: null } }
        '502': { $ref: '#/components/responses/OpenAIUpstream' }
...
components:
  securitySchemes:
    session:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 7-day session from `POST /auth/verify`.
    apiKey:
      type: http
      scheme: bearer
      description: '`mesh_sk_…` from `POST /keys`. Works wherever an OpenAI key works.'
    nodeToken:
      type: http
      scheme: bearer
      description: '`mesh_node_…` from `POST /nodes/register`.'
...
  responses:
...
    OpenAIUnauthorized:
      description: '`invalid_api_key`. Create one with `POST /keys`.'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/OpenAIError' }
          example: { error: { message: 'Invalid API key. Create one with POST /keys.', type: invalid_request_error, code: invalid_api_key, param: null } }
    OpenAIRateLimited:
      description: Per-key rate limit (120/min).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/OpenAIError' }
    OpenAIUpstream:
      description: The upstream provider failed or timed out. You were not charged.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/OpenAIError' }
          example: { error: { message: 'Upstream (openrouter) timed out after 60000ms. You were not charged.', type: upstream_error, code: upstream_timeout, param: null } }
  schemas:
    UnixSeconds:
      type: integer
      description: Unix time in seconds (UTC).
...
    OpenAIError:
      type: object
      description: Error shape under `/v1`, parseable by OpenAI SDKs.
      required: [error]
      properties:
        error:
          type: object
          required: [message, type]
          properties:
            message: { type: string }
            type: { type: string }
            code: { type: [string, 'null'] }
            param: { type: [string, 'null'] }
...
    PrivacyTier:
      type: string
      enum: [trusted, network, upstream_zdr]
      description: Request privacy tier (docs/PRIVACY.md).
...
    PricePerM:
      type: object
      required: [promptUsdPerM, completionUsdPerM]
      properties:
        promptUsdPerM: { type: number, description: USD per 1M prompt tokens. }
        completionUsdPerM: { type: number, description: USD per 1M completion tokens. }
    Model:
      type: object
      required: [id, object, served, listPrice, meshPrice, privacy, mesh_network]
      properties:
        id: { type: string, description: OpenRouter-style id (`vendor/model`) or a short network alias. }
        object: { type: string, const: model }
        created: { type: integer }
        owned_by: { type: string, description: '`mesh` for network models, else the vendor.' }
        name: { type: string, description: Human name; the same string as `displayName`. }
        displayName: { type: string }
        vendor: { type: string }
        tier: { type: [string, 'null'], enum: [frontier, fast, open, null], description: '`frontier` strongest closed models, `fast` cheaper closed models, `open` open weights.' }
        served: { type: string, enum: [network, upstream, both], description: Where the request can land; `both` is Mesh nodes first with upstream fallback. }
        listPrice: { $ref: '#/components/schemas/PricePerM', description: OpenRouter list price. }
        meshPrice: { $ref: '#/components/schemas/PricePerM', description: 'What Mesh bills; the flat network price on a node, else list plus the configured markup (6 % as shipped) or minus a discount.' }
        privacy: { type: string, enum: [network, upstream_zdr], description: Where the prompt is processed. }
        online: { type: integer, description: Nodes advertising this model right now (0 for upstream-only models). }
        guestAllowed: { type: boolean, description: A homepage guest may pick it. }
        mesh_network: { type: boolean, description: Back-compat; true when Macs in the Mesh network can serve it at the network price. }
      additionalProperties: true
    ChatMessage:
      type: object
      required: [role, content]
      properties:
        role: { type: string, enum: [system, user, assistant, tool] }
        content: { type: string }
    ChatRequest:
      type: object
      required: [model, messages]
      properties:
        model: { type: string }
        messages: { type: array, minItems: 1, items: { $ref: '#/components/schemas/ChatMessage' } }
        stream: { type: boolean, default: false }
        stream_options:
          type: object
          properties:
            include_usage: { type: boolean }
        max_tokens: { type: integer, description: Defaults to 1024 on network-served requests. }
        temperature: { type: number }
        top_p: { type: number }
        top_k: { type: integer }
        stop: { oneOf: [{ type: string }, { type: array, items: { type: string } }] }
        seed: { type: integer }
        presence_penalty: { type: number }
        frequency_penalty: { type: number }
        response_format: { type: object }
      additionalProperties: true
    Usage:
      type: object
      properties:
        prompt_tokens: { type: integer }
        completion_tokens: { type: integer }
        total_tokens: { type: integer }
        cost: { type: number, description: 'USD charged to your wallet for this request: the flat network price on a node, list plus the markup (or minus the discount) upstream. On streams it is in the usage chunk and repeated in the final `mesh` chunk.' }
    MeshRoute:
      type: object
      description: How the reply was served. Node fields are present when a Mesh node served it.
      properties:
        route: { type: string, description: '`node`, or the upstream name (`openrouter`).' }
        privacy: { $ref: '#/components/schemas/PrivacyTier' }
        servedBy: { type: string, enum: [your node, trusted node, network node, upstream (ZDR), upstream] }
        nodeId: { type: string }
        chip: { type: string }
        listCostUsd: { type: number, description: 'What the upstream would have charged at list (node-served replies), or did charge Mesh at list (upstream-served replies billed with a markup or discount).' }
        savedUsd: { type: number, description: listCostUsd minus what you paid, never negative. }
    ChatCompletion:
      type: object
      properties:
        id: { type: string }
        object: { type: string, const: chat.completion }
        created: { $ref: '#/components/schemas/UnixSeconds' }
        model: { type: string }
        choices:
          type: array
          items:
            type: object
            properties:
              index: { type: integer }
              message: { $ref: '#/components/schemas/ChatMessage' }
              finish_reason: { type: [string, 'null'] }
        usage: { $ref: '#/components/schemas/Usage' }
        mesh: { $ref: '#/components/schemas/MeshRoute' }
      additionalProperties: true
...